+++ src/main/java/cn/iqianye/mc/zmusic/api/bossbar/BarColor.java
@@ -1,0 +1,12 @@
+package cn.iqianye.mc.zmusic.api.bossbar;
+
+public enum BarColor {
+    PINK,
+    BLUE,
+    RED,
+    GREEN,
+    YELLOW,
+    PURPLE,
+    WHITE,
+    SOLID;
+}