+++ src/main/java/cn/iqianye/mc/zmusic/command/CmdBC.java
@@ -4,11 +4,14 @@ import cn.iqianye.mc.zmusic.ZMusic;
 import net.md_5.bungee.api.CommandSender;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
 import net.md_5.bungee.api.plugin.Command;
+import net.md_5.bungee.api.plugin.TabExecutor;
 
-public class CmdBC extends Command {
+import java.util.ArrayList;
+
+public class CmdBC extends Command implements TabExecutor {
 
     public CmdBC() {
-        super("zm");
+        super("zm", null, "zmusic", "music");
     }
 
     @Override
@@ -17,6 +20,15 @@ public class CmdBC extends Command {
             Cmd.cmd(commandSender, strings);
         } else {
             ZMusic.log.sendErrorMessage("控制台不支持使用任何指令");
+        }
+    }
+
+    @Override
+    public Iterable<String> onTabComplete(CommandSender commandSender, String[] strings) {
+        if (commandSender instanceof ProxiedPlayer) {
+            return Cmd.tab(commandSender, strings);
+        } else {
+            return new ArrayList<>();
         }
     }
 }