+++ src/main/java/cn/iqianye/mc/zmusic/other/Val.java
@@ -2,7 +2,7 @@ package cn.iqianye.mc.zmusic.other;
 
 public class Val {
     public static String thisVer;
-    public static int thisVerCode = 202009150;
+    public static int thisVerCode = 202009170;
     public static boolean isLatest = true;
     public static String updateLog;
     public static String latestVer;